Only run world load handler once and only on the server
WorldEvent.Load is emitted both on client and on server, so that would lead to the mod creating its backpack save dir in the minecraft root dir when joining a server.

Additionally, WorldEvent.Load is emitted for each dimension, but we only care about running one time when a save is loaded so add a check for dimensionId. I wasn't able to find a more elegant solution.